requirements.txt
rdkit-pypi
setuptools
numpy==1.26.4